services/mcp-manager/pyproject.toml
[project]
name = "mcp-manager"
version = "0.1.0"
description = "MCP manager service for the NGEN platform"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "ngen-common",
    "nats-py>=2.0",
    "httpx>=0.27",
    "fastapi>=0.111,<1.0",
    "uvicorn>=0.30,<1.0",
    "ddgs>=7.0",
    "pdfplumber>=0.10",
    "python-docx>=1.0",
    "python-multipart>=0.0.6",
    "landingai-ade>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_manager"]

[tool.uv.sources]
ngen-common = { workspace = true }
